refactored GeoPinning logic to separate from Jersey (gradle 5 had an issue with unnecessary transitive dependencies when using the new "implementation" configuration)
